Finish encoding and output recipe + fasta files
Encoding is done on pixel matrix instead of pixel array to include the
padded pixels (which allows reconstruction of the image, although
this is not required at the moment). Currently, the unhandled cases of encoding are now handled using an
"N" base as a placeholder. Program outputs recipe for decoding and the fasta file including the
DNA sequence.